Allow the user sidebar partial to be used outside the Solidus engine
This should be a no-op for a stock solidus store. Where it comes in
handy is when you are trying to keep a separation of concerns between
custom code vs. Spree code while still maintaining a consistent UI for
the user.

My custom code operates in it's own context (the app or independent engines)
but I have data hanging off of the user model (i.e. info related to the user).
I don't want the user to have multiple admin screens so my custom code has
a light bit of integration to make it fit in with the Spree admin interface
(uses the same layout and uses shared partials such as this sidebar).

This setup generally works as many of the links already explicitly reference
the `spree` engine. For example the main menu does this at:

https://github.com/solidusio/solidus/blob/master/backend/app/helpers/spree/admin/navigation_helper.rb#L17

This commit updates this sidebar partial to follow the pattern set by the
main menu and explicitly indicate the engine. This allows it to work within the
spree engine as well as outside the spree engine.