Add Variant Pricer and related Pricing Options class
The variant pricer class can either be instantiated using a Hash or using an order.
When setting the price for a line item and when displaying the price
of a variant, so far we're assuming that only the currency is relevant
for the price, when really it could be a number of other things.

Calculating a line item price is substantially different from showing
the price for a variant in a frontend helper, as in one case we use
the `current_currency` variable, while in the other we take things from
the order. This class provides a single source of truth for all things
necessary to choose a price in both cases.

Added tests to reflect that the option keys have to reflect columns
on the Spree::Price model, as it is going to be used for ActiveRecord queries.

Also added a #cache_key and a #currency method to replace the differnt
uses of `current_currency` in the frontend.