Remove the required attribute from shipping category
This attribute works great on non select2 elements. Unfortunately, our
version of select2 makes the browser error message appear at the top of
the form, nowhere near the actual dropdown. This is confusing because
the administrator does not know what field the validation is
referencing.

This commit simply removes the front-end validation for this. If an
admin creates a product without a shipping category selected the error
message will be displayed after a page load.