Stop expecting the taxation system to adjust shipment amounts
Once we simplify the taxation system to trust that prices are correct
from the get-go, we can not expect it to change shipment amounts.

This should also not be necessary as the shipping method system already
allows for stratified prices by zone.

Essentially, these expectations add an unnecessary implementation burden.