Only run the VatPriceGenerator when absolutely necessary
Rather than using a hard-to-understand callback that would always
create prices whenever you update the backend price, this commit
introduces a checkbox in both the variant and product form that will
update VAT prices for that product if you tick the checkbox.

The only exception to this is new products/variants, which will always
run the VatPriceGenerator before validation.