Remove clear cache button from admin
Our existing "General Settings" page has a "Clear cache" button.

On stores I've worked on I wouldn't want this exposed to admins. We
should (and afaik we do) expire caches every time the underlying DB
object is changed.

This is motivated partly because I want to replace the
GeneralSettingsController with a StoreController, and this is the only
thing on that page that doesn't directly relate to a store.