Payment Factory should associate source with User
In the process of creating a Payment via the factories, a new CreditCard
record is created, but not associated with any user.

If we create a Payment via one of the Order factories, this can be quite
confusing, as the created Order may have a payment, but the created
Credit Card associated with the payment is not associated with the User
that created the order.

To correct this, we can provide the order's associated user to the
credit card we create as the payment source.