Merge tax updating code and deprecate Order#create_tax_charge!
Previously we had two different sets of tax updating code.  One was
Order#create_tax_charge! and the other was the code in OrderUpdater.
The effect of the former encompasses the effect of the latter, so
functionally we can merge these.

I will address the increased load (`create_tax_charge!` is more
expensive) in an later commit.